pyproject.toml
[project]
name = "cross-pictures"
version = "0.1.0"
description = ""
authors = [
    {name = "AnaSofia94",email = "anasofiamoreira94@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry]
packages = [{include = "cross_pictures", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "0.122.0"
uvicorn = "0.38.0"
environs = "^14.5.0"
pytest = "^9.0.2"
httpx = "^0.28.1"
onnxruntime = "1.20.0"
numpy = "^2.4.1"
pillow = "^12.1.0"
python-multipart = "^0.0.21"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
